gateway/requirements.txt
fastapi
uvicorn
httpx
tenacity

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi

